requirements.txt
fastapi
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]